+++ src/clj/py99/routes/services.clj
@@ -80,6 +80,7 @@
    (-> (db/points? {:login login})
        (select-keys [:login :wil :py99 :comm :m1 :m2 :m3 :e1]))))
 
+;; update は grading の仕事．
 (defn py99 [{{:keys [login]} :path-params}]
   (response/ok
    {:login login
@@ -119,12 +120,15 @@
 (comment
   (db/update-goal! {:login "mijuhashi" :pt 10})
   :rcf)
-;; (defn seven-four! [{{:keys [login]} :path-params :as request}]
-;;   (let [pt (-> (get-in request [:params :pt]) Intger/ParseInt)]
-;;     (log/debug "seven-four! pt" pt)
-;;     (db/update-seven-four! {:login login :pt pt})
-;;     (response/ok {:login login
-;;                   :pt pt})))
+
+(defn seven-four! [{{:keys [login]} :path-params :as request}]
+  (let [pt (-> (get-in request [:params :pt])
+               Integer/parseInt)]
+    (log/debug "seven-four! pt" pt)
+    (if (= 1 (db/update-seven-four! {:login login :pt pt}))
+      (response/ok {:login login
+                    :pt pt})
+      (response/ok {:error "can not update"}))))
 
 (defn service-routes
   []
@@ -141,5 +145,4 @@
    ["/py99/:login/:pt" {:post py99!}]
    ["/comm/:login/:pt" {:post comm!}]
    ["/goal-in/:login" {:get goal-in :post goal-in!}]
-   ;;["/seven-four/:login" {:post seven-four!}]
-   ])
+   ["/seven-four/:login" {:post seven-four!}]])