+++ CHANGELOG.md
@@ -38,10 +38,12 @@ https://forums.docker.com/t/docker-desktop-shows-image-has-a-python-wheel-vulner
 
 
 ## 0.90-SNAPSHOT
+アップデートは grading の仕事．
 - GET /api/goal/:login
 - POST /api/goal/:login
 - db/update は戻り値をチェックした方がいい．
 - `alter table gradings rename e6 to seven_four;`
+- POST /seven-four/:login
 
 ## 0.89.861 / 2024-02-18
 - POST /api/py99/:login/:pt